- renamed id to externalId in UserBankProfile and BankAccount - added ExchangeRates to the DB - added id to the all entities in DB - uncommented fatalError in CoreDataStack

+++ Accountant/Model/Exchange Structures/National Bank/CurrencyExhangeNB.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct CurrencyExhangeNB: Codable {
-    let code : Int
+    let code : Int16
     let txt : String
     let rate : Double
     let currency : String
